cmd/compile/internal/ssa: allow jumps to clobber flags
This is the minimal change to allow SSA to use the two-register Bxx
instructions in RISC-V.  We are going to use a COMPARE
pseudo-instruction which allows the two registers to be bundled as a
single "flags" argument.

Normal flag registers are treated as an implicit Phi; flagalloc can
either rematerialize a flag in the basic block that uses it, or in all
predecessors of the block.  However because our branch instructions
steal the register numbers from the fake COMPARE, there has to be
exactly one COMPARE value; which effectively means that it has to be
rematerialized in the basic block that uses it.

This patch adds an arch config option jumpsSetFlags which forces
flagalloc to ensure that flags are dead at the start of all basic
blocks.

Change-Id: I594a13fda2cdd0ca35341c1ac842f22065abb47e