cmd/compile: Use two-register Bxx on RISC-V
The core SSA code assumes that each block has exactly one control value.
Rather than changing that, we instead allow a single value of type Flags
to stand-in for the pair of registers that are going to be used in the
branch; for this to work, all instructions that write registers have to
clobber the flags, and the flags cannot be assumed live across basic
blocks (because if they were set in two different predecessors, regalloc
might have used different register numbers in each predecessor).

Change-Id: I4de6a465baaa010bf5f5d7f6d68ce7634ce7c663