cmd: Add and use direct branches for RISC-V
RISC-V has a one-instruction branch JAL with a 1 MiB range, and a
two-instruction sequence with a 2 GiB range.  Currently we always use
the two-instruction sequence.  This patch adds support for the JAL
instruction and adds linker support for trampoline generation.
Functions larger than 256 KiB continue to use the two instruction
sequence because the trampolines might not be reachable, and deferreturn
uses a two-instruction sequence regardless of function size to keep the
runtime code simple.

The ELF RISC-V toolchain does not use trampolines, instead it generates
the two-instruction sequence in the compiler and has a linker pass that
shortens shortenable branches.  This could also be adopted for Go,
alongside trampoline support, but it would require a significant amount
of new linker machinery to support within-function relocations (since
the compiler no longer knows the offsets of conditional branches) and
pclntab rewriting for the new PC offsets.

Change-Id: I47d03312b5c82a042378b91587d71cf043324c06